Use a precedence group for Decode operators
We're defining this in relation to the operators defined in Runes.
Specifically, we want this to have a higher precedence than any of those
operators. That means dropping this just above ApplicativeSequence.

Note that because of the precedence change wrt <|>, we now need to add parens
to get it all to compile properly.